+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed gravestone block can't explode
+- Updated to 1.14.4